Docker/AlertManagerDocker/requirements.txt
streamlit==1.22.0
kafka-python==2.0.2
